refactor(domain): LottoGenerator 전략 인터페이스 추가
