+++ src/main/java/lotto/domain/Lotto.java
@@ -15,5 +15,4 @@ public class Lotto {
             throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
         }
     }
-
 }
