+++ src/main/java/lotto/domain/RandomLottoGenerator.java
@@ -1,0 +1,11 @@
+package lotto.domain;
+
+import camp.nextstep.edu.missionutils.Randoms;
+import lotto.strategy.LottoGenerator;
+
+public class RandomLottoGenerator implements LottoGenerator {
+    @Override
+    public Lotto generate() {
+        return new Lotto(Randoms.pickUniqueNumbersInRange(1, 45, 6));
+    }
+}
