+++ src/main/java/lotto/strategy/LottoGenerator.java
@@ -1,0 +1,7 @@
+package lotto.strategy;
+
+import lotto.domain.Lotto;
+
+public interface LottoGenerator {
+    Lotto generate();
+}
